+++ test_project/test_project/settings/base.py
@@ -39,6 +39,7 @@ INSTALLED_APPS = [
     "django.contrib.messages",
     "django.contrib.staticfiles",
     "django_extensions",
+    "rest_framework",
     "django_dramatiq",
     "django_ctb",
     "test_project",